Restructure front page of docs
This preserves basically all the previous information (modulo fixing a few
typos, and some minor rephrasing for clarity), but breaks it up into a series
of sub-pages and includes them in the toctree.

The idea is that readers aren't hit with a wall of text on first hitting the
docs, but instead get a short introduction and then can quickly see what's
available in the toc.