Clean up introductory material
Primarily, assert that it is outside the scope of this documentation to
clarify astrophysics, and therefore remove the somewhat confusing tutorial
material. This is replaced with a link to the Huppenkothen et al paper, which
is not so introductory but at least provides some references the interested
user can follow.

I do not suggest trying to reinstate a from-the-ground-up tutorial in the
Stingray docs (unless there are significant resources available to make it a
really good one!), but adding further links to introductory material where
available seems like a good idea.