Added code to read unit wise data of LAXPC
There is a quick fix being implemented at the moment.  LAXPC's current software results in a column heading called "LAXPC_No." The current implementation of stingray converts this to an attribute named laxpc_no., which leads to an error in the compilation. For testing purposes, the original event file was modified to have a different name for the column but this needs to be changed in the stingray software.