Commit to read the LAXPC files as is.
Changes in the events.py
Removed the functions specific to LAXPC and modified a filtering function to work on detector_id attribute instead.

Changes in missions.py
If the mission is AstroSat then the detector ID columns maps to LAXPC_No.

Changes in io.py
If the code detected a "." in the attribute name, it is replace with an underscore.