Updated the tests for the LAXPC spec code
Changes to test_io.py
Added a check to see is the instrument is LAXPC3

Changes to test_events.py
Added a test to read the LAXPC file. This also catches warnings present while reading the LAXPC files.

Changes to missions.py
Manual assignment of the Energy/PI column and Instrument keyword from the header of the event file.